Fix final renames
