+++ src/pages/photos/_id.vue
@@ -30,7 +30,7 @@
 <script>
 import axios from 'axios'
 import { mapActions, mapMutations, mapState } from 'vuex'
-import featureFlags from '~/featureFlags'
+import featureFlags from '~/feature-flags'
 import { FETCH_IMAGE, FETCH_RELATED_IMAGES } from '~/store-modules/action-types'
 import { SET_IMAGE, SET_RELATED_IMAGES } from '~/store-modules/mutation-types'
 import iframeHeight from '~/mixins/iframe-height'
